chore: refactor credentials loading
For the migration to the DBT platform use new env vars if present:
REDIS_ENDPOINT
DATABASE_CREDENTIALS
